+++ components/eamxx/src/physics/rrtmgp/atmosphere_radiation.hpp
@@ -42,7 +42,7 @@ public:
 public:
   // The three main interfaces for the subcomponent
   void initialize_impl (const RunType run_type);
-  void run_impl        (const int dt);
+  void run_impl        (const double dt);
   void finalize_impl   ();
 
   // Keep track of number of columns and levels
